[gallery] Don't use opengl graphicssystem or QGLWidget viewports
The opengl graphicssystem is simply wrong; we default to the meego
graphicssystem, which switches between opengl and raster as needed.
Using opengl directly will cause the process to hold contexts even
when minimized, which wastes huge amounts of memory.

QGLWidget as a viewport is also mostly redundant - GL is used for
everything under the meego graphicssystem. It also causes excessive
memory usage.