[jolla-gallery] Handle video content touches as control commands. Contributes to JB#28599
Previously, playback of the video was only started if the user
explicitly pressed the play button.  This commit adds handling
of touches to anywhere in the video poster as control commands
(that is, either close splitview and begin or resume playback,
or pause playback and open splitview).

Contributes to JB#28599