Make cover photo's frames look ok.
When scaling down, we need to use mipmapping or area sampling to avoid
artifacts in high-contrast areas, such as the antialiased edge on the
photo frame. As mipmapping is not available for Image on 5.2, we have
to use area sampling (which downscaling through sourceSize does). This
is also potentially a memory saver as the image is not used in its
full height.