[jolla-gallery] Return to previous state after minimize/restore. Contributes to JB#28599
This commit ensures that if the user is playing a video in fullscreen
in gallery, and then minimize and then restore, the gallery goes to
fullscreen view and restarts playback.  It also ensures that if the
user is in split-view mode and they then minimize and then restore,
that the gallery remains in split-view mode.

Contributes to JB#28599