[ambience] Adapt to the decoupling of images and ambiences. Contributes to JB#15647
Use Ambience.create callback API to receive a notification and id when a
requested ambience has been created instead of monitoring an ambience
model.  The images for ambiences are properly deleted when the ambience
as well now, so the ability to remove an ambience while it is active has
been locked down.  With that in mind don't activate created ambiences until
the edit dialog is closed so that there is an initial opportunity to remove
it.