[gallery] Sort images by date taken. Fixes JB#10645
Sorting was changed long time ago on commit 102f8824560e2
without being mentioned or rationale provided.

Changed to date taken. Helps the case when a set of files gets transferred
to the device. Order in gallery is no longer depending that much
on transfer order, but rather how old the pictures are.