[gallery] Fix videos losing their position when app on background. JB#60729
Application put to background was resetting the source as empty string
and then returning it on foreground, resulting in position being lost.

Looks like regression on commit cac549c4c6429. The 'active' property
should be set false as it controls the media keys, but the source
file shouldn't be reset.