[gallery] Avoid cover warning on assigning undefined to double. JB#37217
Especially if a new image appears on the model, the delegate can
get index 8 while the arrays on function layout() are having
information for maximum index 7, and thus returning an empty array.
Proteting this case now by just moving the exceeding item to bottom
of the photo stack.

Adjusted some coding convention details while at it.